Return full taps and signal for selected modes
When we use the apply keyword in equalise_signal when equalising over some selected modes only, we return the full filter taps (with some not being converged) and the signal only for the selected modes. It's not sure if this is the best way to do it, because it results in different behaviour when we apply the taps separately.