Move symbol decision after the synchronization for BER and SER calc
The calculations of BER and SER where doing the symbol decisions before the synchronization of Tx and Rx. This might improve the sync performance, but it also makes a lot of other things trickier. Considering that the GMI calculations were already doing it after it's unlikely that this change will cause issues.